research_assistant/requirements.txt
# =========================
# Core Django
# =========================
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
gunicorn
whitenoise

# =========================
# Database (Postgres)
# =========================
psycopg2-binary==2.9.11

# =========================
# Environment
# =========================
python-dotenv==1.2.1

# =========================
# AI APIs
# =========================
groq==1.0.0
google-genai

# =========================
# RAG / Vector Search
# =========================
faiss-cpu==1.13.2
numpy>=1.26.0

# =========================
# PDF Processing
# =========================
PyMuPDF==1.27.1
pillow>=10.0.0

# =========================
# HTTP
# =========================
requests>=2.31.0
httpx>=0.27.0

pytesseract
# =========================
# Optional (ONLY if using Celery)
# =========================
celery==5.6.2
redis>=5.0.0

